requirements.txt
selenium
undetected-chromedriver
pandas
requests
tqdm
tenacity
pillow

google-generativeai
google-cloud-storage
google-api-core
vertexai